Improve logic for allowing access to Verify pages
**Why**:
- Users who have already completed identity verification should not
be able to directly access the `/verify/cancel`, `/verify/fail`, and
`/verify/retry` pages because those are only meant for people who have
not verified yet.
- Users who have not already been verified should not be able to access
the `/verify/activated` page because that is only meant for those who
already have an active profile.
- Only users who have exceeded the max allowable proofing attempts
should be allowed to access the `/verify/fail` page.
- The "Back" link on the `/verify/activated` page that points to the
`/verify` endpoint doesn't make sense because it will send right back
to the `/verify/activated` page. I removed the link because the only
place where it makes sense to link back to is the profile page, which
is already accessible via the "My account" link or the login.gov logo.